microservice/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
google-genai>=1.69.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
